Fix the inconsistency during preprocessing
There are a few inconsistencies between load_graph and
nodes_preparation.
In nodes_preparation originally, it reads the files right from
the `Graph` or `Graph/Pickle` directory.
In the meanwhile, load_graph stores files under `Graph/<location>`.
In this case, it's `Graph/Toy_Example`.
Since in the given example, the files are also stored in
`Graph/<location>`, I believed that these files are intended to
be classified by their location, which means following the way
load_graph does.
Thus, some changes in nodes_preparation are conducted to read
files based on the corresponding location name.
Though it still can't be executed with the given inadequate
example files, it provides better clarity and consistency now.