Check if the logging directory exists before training
In the original reinforcement training program, the monitor is set
to the logging directory without checking whether it exists or not.
Therefore, a typical "no such file or directory" error occurs when
users first try to run it after cloning the repo.
The program now will try to create the directory, errors are
expected to occur only when the creation fails.