Check if the image directory exists before visualizing
In the original visualization program, the result is visualized as
a image in the image directory without checking whether it exists
or not.
Therefore, a typical "no such file or directory" error occurs when
users first try to run it after cloning the repo.
The program now will try to create the directory, errors are
expected to occur only when the creation fails.