Remove redundant directories for reinforcement learning
The logging directory should be created before the monitor is set,
and it happens before the callback can take place.
Therefore, there is no need to do that again in the callback, thus
removed.
Moreover, during the process, there is nothing in the directories
created when a better model shows up. The model is saved as a zip
file, so no need to create new directories for that.